Quote arguments with whitespace when printing command
Currently, printing baked commands doesn't represent the way arguments
are actually passed down to the program when values contain spaces.

The solution is to take such values in single quotes.

This is primarily to be able to copy printed commands
and run them in a shell.

BEFORE:

```python
>>> from sh import ls
>>> print(ls.bake('How I Met Your Mother'))
/bin/ls How I Met Your Mother
```

AFTER:

```python
>>> print(ls.bake('How I Met Your Mother'))
/bin/ls 'How I Met Your Mother'
```